requirements.txt
ultralytics
google-generativeai
python-dotenv
opencv-python

